fix: run vm script for (reverse) fqdn host names
the name of the script is only construed from the RFC1035 host label,
not the entire (reverse) fqdn host name. Hence `$HOST`, when it's a
(reverse) fqdn identifier can produce a mismatch.

Since we always build one host within the context of this script, as a
work around, we can safely use globbing, here.